Fixed segfault on show

+++ production/main.cpp
@@ -108,7 +108,7 @@ int main(int argc, char *argv[]) {
     // Start threads.
     pthread_t *threads = new pthread_t[num_threads];
 
-    for (int i = 0; i < num_threads; i++) {
+    for (size_t i = 0; i < num_threads; i++) {
         pthread_create(threads + i, NULL, runWorker, NULL);
     }
 
@@ -142,11 +142,11 @@ int main(int argc, char *argv[]) {
     }
 
     // Force all threads to exit
-    for (int i = 0; i < num_threads; i++) {
+    for (size_t i = 0; i < num_threads; i++) {
         tasks -> push(nullptr);
     }
 
-    for (int i = 0; i < num_threads; i++) {
+    for (size_t i = 0; i < num_threads; i++) {
         pthread_join(threads[i], NULL);
     }
